Switch to {lint-staged} for running pre-commit hook
Using {lint-staged} instead of running raw commands in the pre-commit git
hook fixes a problem and provides us a benefit.

The problem is that, the auto-fixing feature of eslint and prettier
conflict with the commit hook. Suppose we run {git commit} and there is
a prettier/eslint issue in file {a.js}. Prettier/Eslint will fix the issue
in {a.js} and the commit is applied to git, while the file {a.js} which
is changed by the tool is not staged and so not committed. In this scenario
the developer should stage {a.js} again and add a new commit or amend the
previous one, which is not the ideal scenario for sure. {lint-staged}
fixes this issue by applying the result of auto-fixing to the current
commit without the need of an explicit staging.

In addition to this problem, {lint-staged} only runs the commands on
staged files and not the whole project, clearly causing the hook to run
much faster.