Fix issue in `typeorm:generate` npm script
When `typeorm:generate` is run, it uses `ts-node` internally. By default,
if we try to import modules without extension for file paths, `ts-node`
raises errors because of missing extensions. To prevent this, we need to
provide an experimental flag to `ts-node` like we have done for `start`
npm script.

Fix the error by providing the flag to `typeorm-ts-node-esm` in `typeorm`
npm script, which is used by `typeorm:generate`.