Fix null prototype objects logging
There is an issue in winston which causes it to throw an error when a
null-prototype object is passed to logger expressions as `meta.message`.
Because all `JsonBig.parse` returned objects are null-prototype, we get
the error wherever we pass them as `meta.message` to logger expressions.

Fix the issue by applying a patch until it is fixed by winston.

Link to GitHub issue:
https://github.com/winstonjs/winston/issues/2267