Add a new json parser to utils
The problem with json parser of `json-bigint` is that, you cannot be explicit
about the path of object props you want to convert to bigint. The current
parser converts all numbers in the json to bigint, which may cause many
issues.

Implement a new parser which only convert paths to bigint, which are
explicitly set through the second parameter of the parser.