Fix trigger event change box creation fee issue
When creating a trigger, if the payment box used as fee has 1. exactly
the same value as one fee and 2. has some token, then when we try to create
the change box for the tokens, there is not enough erg to use for change
box creation fee, and we get a `BoxCreationError`.

Fix the issue by doubling the required fee value.