Fix reveal duplicate fee boxes issue
When creating triggers, if the transactions are queued, there is a possible
scenario in which we get two same boxes to spend for fees. This causes a
`duplicate inputs found` error in tx builder.

To prevent so, reject all duplicate boxes when selecting them in
`getUserPaymentBox`.

In addition, add type for an array in `getUserPaymentBox`.