Fix tx inputs checking issue
When calling `checkTxInputs`, there is a case in which the inputs may be
spent, but in the same tx. For example, we may submit the tx, but because
of some network error, we think the tx is not submitted and later, when
we check the inputs, we see that they are spent and the method returns
false, but that's not the expected result.

Fix the issue by adding a new `txId` parameter to the method to make sure
inputs are not spent in the same tx.

In addition, change some interface and method names.