Refactor commitment redeem job
Commitment redeem is not related to observations since there might be
more than one commitment related to the same observation, and there
might be some invalid commitments that we don't have the corresponding
observation in the database. Then after creating the redeem transaction
we don't need to update the observation status.

Commitment redeem should be done if after a specified time (in blocks),
the commitment has one of these status:
1- Not triggered after the specified period
2- Created after the related trigger
3- It's a duplicate commitment and a valid one merged to create the
trigger (WID exists in trigger)