Use inputs max height in case of api problem
To prevent the transaction outputs to have smaller height than the
inputs we used the maximum of network height and inputs' heights.
EIP-39 prevents transactions to have smaller output height. We had
a bug report that in some transactions outputs have creation height set
to 0, and the transaction is not accepted by the nodes. It might be the
node API problem where we get the latest network height. Then we added
this update to prevent a wrong height for boxes.